Rename mailbox lifecycle hook
