+++ ui/src/taskpane/hooks/useTaskPaneController.ts
@@ -6,12 +6,12 @@ import {
   PersistedTaskPaneState,
 } from "../helpers/outlook-persistence";
 import { resolveStorageKeyForCurrentItem } from "../helpers/outlook-mailboxItem";
-import { registerTaskpaneVisibilityHandler } from "../helpers/outlook-runtime";
 import { cancelSendOperation, clearSendOperation } from "../helpers/outlook-runtimeLogic";
 import { copyTextToClipboard } from "../helpers/clipboard";
 import { insertResponseIntoBody } from "../helpers/emailBodyInsertion";
 import { useTaskPaneStatePersistence } from "./useTaskPaneStatePersistence";
 import { useSendLifecycleHandler } from "./useSendLifecycleHandler";
+import { useMailboxLifecycleHandler } from "./useMailboxLifecycleHandler";
 import { describeError } from "../utils/outlook-errorHandling";
 
 export interface TaskPaneActions {
@@ -98,54 +98,7 @@ export const useTaskPaneController = (): TaskPaneController => {
     }
   }, [resumePendingOperationIfNeeded]);
 
-  React.useEffect(() => {
-    isMountedRef.current = true;
-    console.info("[Taskpane] Task pane mounted. Initializing lifecycle handlers.");
-
-    const initialize = async () => {
-      await refreshFromCurrentItem();
-      visibilityCleanupRef.current =
-        await registerTaskpaneVisibilityHandler(refreshFromCurrentItem);
-    };
-
-    void initialize();
-
-    const mailbox = Office.context.mailbox;
-    const itemChangedHandler = () => {
-      console.info("[Taskpane] Office item changed event received. Triggering refresh.");
-      void refreshFromCurrentItem();
-    };
-
-    if (mailbox?.addHandlerAsync) {
-      mailbox.addHandlerAsync(Office.EventType.ItemChanged, itemChangedHandler, (result) => {
-        if (result.status !== Office.AsyncResultStatus.Succeeded) {
-          console.warn("[Taskpane] Failed to register ItemChanged handler.", result.error);
-        } else {
-          console.info("[Taskpane] ItemChanged handler registered.");
-        }
-      });
-    }
-
-    return () => {
-      isMountedRef.current = false;
-      console.info("[Taskpane] Task pane unmounted. Cleaning up handlers.");
-
-      if (visibilityCleanupRef.current) {
-        void visibilityCleanupRef.current();
-        visibilityCleanupRef.current = null;
-      }
-
-      if (mailbox?.removeHandlerAsync) {
-        mailbox.removeHandlerAsync(Office.EventType.ItemChanged, (result) => {
-          if (result.status !== Office.AsyncResultStatus.Succeeded) {
-            console.warn("[Taskpane] Failed to remove ItemChanged handler.", result.error);
-          } else {
-            console.info("[Taskpane] ItemChanged handler removed.");
-          }
-        });
-      }
-    };
-  }, [refreshFromCurrentItem]);
+  useMailboxLifecycleHandler({ refreshFromCurrentItem, isMountedRef, visibilityCleanupRef });
 
   const updateOptionalPrompt = React.useCallback(
     (value: string) => {
